requirements.txt
alembic==1.5.5
appnope==0.1.2
backcall==0.2.0
blinker==1.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
decorator==5.0.9
Flask==1.1.2
Flask-Admin==1.5.8
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
Flask-HTTPAuth==4.2.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
ipython==7.24.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib-inline==0.1.2
mysqlclient==2.0.3
oauth2==1.9.0.post1
oauthlib==3.1.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
slack-sdk==3.7.0
SQLAlchemy==1.3.23
stripe==2.58.0
traitlets==5.0.5
urllib3==1.26.3
URLObject==2.4.3
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
